rectified code
