+++ controllers/index.js
@@ -7,6 +7,7 @@ const dashboardRoutes = require('./dashboardRoutes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
-router.use('/signin', dashboardRoutes);
+// router.use('/signin', dashboardRoutes);
+// // router.use('/signup', dashboardRoutes);
 
 module.exports = router;
